+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+- drafts to use versioning
+
+### Changed
+
 - update `searchCivicPlusHCMSContentItems` to call HCMS directly
 
 ### Added